add some javascript, change some codes in Survey.cshtml, added new action UserAnswers
javascript codes(nextButton and prevButton) change Questions in the Survey View and finishButton take selected values and userid and send them to UserAnswers Action with using Ajax.
In Survey.cshtml, i change foreach to for loop and organized to for loop
UserAnswers for the moment take selectedvalues and userid. Next part will be about sending these to API.